Data & sorting data script
Upload of the data in two different folders. HA-SA it's a folder with the data splitted in two different categories. HA: Happy and SA: Sad. 
mixed_data is a folder where there are all the images without being categorized. It has inside a folder with outliers. 
The python script is the script we've used to create the folders. It needs the original data in order to reorder the data in folders